refactor: extract scheduler policy
